Spawn child processes
