+++ utils/exampleExpander.js
@@ -1,19 +1,3 @@
-const uuid = require('uuid/v1')
-
-class Logger {
-  constructor(name) {
-    this.name = name
-    this.id = uuid()
-    this.t = process.hrtime()
-    console.log('timer [%s] started for [%s]', this.id, name)
-  }
-
-  stopAndLog() {
-    const diff = process.hrtime(this.t)
-    console.log('timer [%s] for [%s] took [%s] seconds [%sms]', this.id , this.name, diff[0], diff[1] / 1000000)
-  }
-}
-
 const sortObjectKeys = obj => JSON.stringify(obj, Object.keys(obj).sort())
 
 const removeDuplicates = arr => {
@@ -32,14 +16,12 @@ const expandExamples = examples => {
   let expandedExamples = examples
   let previousLength = -1
   // decrease this in order to prevent stack overflow
-  let remainingRuns = 3
+  let remainingRuns = 1
 
   while(expandExamples.length !== previousLength && remainingRuns-- > 0) {
-    console.log('remaining runs [%s], example length [%s]', remainingRuns, expandedExamples.length)
+    // console.log('remaining runs [%s], example length [%s]', remainingRuns, expandedExamples.length)
     previousLength = expandedExamples.length
-    const t = new Logger(`expand ${expandedExamples.length} items`)
     expandedExamples = expand(expandedExamples)
-    t.stopAndLog()
   }
 
   return expandedExamples
